Added test echo function.

+++ server/jobs/build_function.sh
@@ -1,10 +1,24 @@
 #!/bin/bash
 
+case "$1" in
+	-u|--update)
+		UPDATE=1;
+		shift
+	;;
+	-n|--new)
+		NEW=1;
+		shift
+	;;
+	*)
+	break
+	;;
+esac
+
 FUNCTION_NAME=$1
 
 # build the function
 compile() {
-	cd $1 && GOOS=linux go build -o main
+	cd $FUNCTION_NAME && GOOS=linux go build -o main
 }
 
 # package the function
@@ -19,10 +33,16 @@ upload() {
 	  --region us-east-1 \
 	  --function-name $1 \
 	  --memory 128 \
-	  --role arn:aws:iam::947945882937:role/lambda_user \
+	  --role arn:aws:iam::016267150191:role/LambdaRole \
 	  --runtime go1.x \
 	  --zip-file fileb://main.zip \
 	  --handler main
+}
+
+update() {
+	aws lambda update-function-code \
+		--function-name $1 \
+		--zip-file fileb://main.zip
 }
 
 echo "Compiling"
@@ -31,8 +51,20 @@ compile $FUNCTION_NAME
 echo "Compressing"
 compress $FUNCTION_NAME
 
-echo "Uploading"
-upload $FUNCTION_NAME
+if ! [[ -z $UPDATE ]]; then
+	echo "Updating"
+	update $FUNCTION_NAME
+elif ! [[ -z $NEW ]]; then
+	echo "Uploading"
+	upload $FUNCTION_NAME
+else
+	echo "NOOP"
+fi
+
+if [[ "$?" = "1" ]]; then
+	echo "Maybe try updating the function with flag --update"
+	exit 1
+fi
 
 echo "Cleaning up"
 rm main
